Fix compiler warnings across Unrar, SystemIV, and Defcon... Deprecated GetVersionEx in Unrar, just wrapped in pragmas as its not important. int/unsigned int conversions everywhere static_cast spammed everywhere as a solution.

+++ contrib/systemIV/contrib/unrar/unrar.cpp
@@ -12,7 +12,7 @@ MemMappedFile::MemMappedFile(char const *_filename, unsigned int _size)
 {
 	m_data = new unsigned char[_size];
 	m_filename = strdup(_filename);
-	unsigned int len = strlen(_filename);
+	unsigned int len = static_cast<unsigned int>(strlen(_filename));
 	for (unsigned int i = 0; i < len; ++i)
 	{
 		if (m_filename[i] == '\\')
